refactor: cleanup unnecessary (#193)
* chore: remove uncessary code


* chore: remove more uncessary code


* chore: more cleanup


---------
