+++ foyer-storage/src/lib.rs
@@ -28,7 +28,6 @@ pub mod catalog;
 pub mod device;
 pub mod error;
 pub mod flusher;
-pub mod flusher_v2;
 pub mod generic;
 pub mod judge;
 pub mod lazy;
